+++ screenpy_examples/screenpy_selenium/github/features/conftest.py
@@ -2,3 +2,4 @@ from screenpy.pacing import the_narrator
 from screenpy_adapter_allure import AllureAdapter
 
 the_narrator.adapters.append(AllureAdapter())
+
