+++ screenpy_examples/screenpy_selenium/github/features/pytest.ini
@@ -1,0 +1,26 @@
+[pytest]
+norecursedirs = .pytest_cache 
+
+python_files = *.py
+;python_files = t*.py 
+; using *.py  is BROKEN in python 3.10.  
+
+
+addopts = --show-capture=no
+          --capture=no
+          --tb=native
+;          --showlocals
+          --verbosity=0
+          -rA
+          --color=yes
+;          --disable-warnings
+
+log_cli = on
+log_level = DEBUG
+log_cli_format = %(fileline)-25s %(msg)s
+;log_cli_format = %(asctime)s %(fileline)-25s %(msg)s
+;log_cli_format = %(asctime)s %(levelname)s %(fileline)-25s %(msg)s
+;log_cli_format = %(msg)s
+
+
+
