+++ screenpy_examples/screenpy_selenium/github/tasks/search_github.py
@@ -9,8 +9,9 @@ from screenpy import Actor
 from screenpy.pacing import beat
 from screenpy_selenium.actions import Enter, Wait
 
-from ..ui.github_header_bar import SEARCH_INPUT
-from ..ui.github_search_results_page import RESULTS_MESSAGE
+from screenpy_examples.screenpy_selenium.github.ui.github_header_bar import SEARCH_INPUT
+from screenpy_examples.screenpy_selenium.github.ui.github_search_results_page import \
+    RESULTS_MESSAGE
 
 
 class SearchGitHub:
