+++ screenpy_examples/screenpy_selenium/the_internet/README.md
@@ -1,0 +1,15 @@
+# The Internet Example
+
+[The Internet](http://the-internet.herokuapp.com/)
+is a website
+created by Dave Haeffner.
+It provides several examples
+of modern web design
+to practice interacting with
+through Selenium.
+
+
+There are two editions of these tests.
+ * [unittest](features_unittest/) 
+ * [pytest](features_pytest/)
+
