+++ screenpy_examples/screenpy_selenium/the_internet/features_pytest/README.md
@@ -1,0 +1,30 @@
+# The Internet Example (pytest edition)
+
+[The Internet](http://the-internet.herokuapp.com/)
+is a website
+created by Dave Haeffner.
+It provides several examples
+of modern web design
+to practice interacting with
+through Selenium.
+
+This example uses [pytest](https://docs.pytest.org/en/7.2.x/)'s test organization style.
+
+## Setup
+
+It is assumed you have `geckodriver` installed.
+
+### Virtual environment
+
+Create a virtual environment for screenpy_examples. See [link-here]() for
+guidance.  _the link should show how to install the entire package_
+
+## Running the Tests
+
+Navigate to the features_pytest folder:
+    
+    cd your_projects/screenpy_examples/the_internet/features_pytest/
+
+Run The tests:
+
+    python -m pytest
