+++ screenpy_examples/screenpy_selenium/the_internet/features_pytest/test_frames.py
@@ -8,7 +8,12 @@ from screenpy.resolutions import ReadsExactly
 from screenpy_selenium.actions import Open, SwitchTo
 from screenpy_selenium.questions import Text
 
-from ..user_interface.iframe import CONTENT_BOX, URL, WYSIWYG_IFRAME
+from screenpy_examples.screenpy_selenium.the_internet.user_interface.iframe import \
+    (
+    CONTENT_BOX,
+    URL,
+    WYSIWYG_IFRAME,
+)
 
 
 class TestFrames:
