+++ screenpy_examples/screenpy_selenium/the_internet/features_unittest/test_alerts.py
@@ -19,7 +19,8 @@ from screenpy_selenium.actions import (
 )
 from screenpy_selenium.questions import Text, TextOfTheAlert
 
-from ..user_interface.javascript_alerts import (
+from screenpy_examples.screenpy_selenium.the_internet.user_interface.javascript_alerts import \
+    (
     JS_ALERT_BUTTON,
     JS_CONFIRM_BUTTON,
     JS_PROMPT_BUTTON,
