+++ screenpy_examples/screenpy_selenium/the_internet/features_unittest/test_drag_and_drop.py
@@ -13,7 +13,8 @@ from screenpy_selenium.abilities import BrowseTheWeb
 from screenpy_selenium.actions import Chain, HoldDown, MoveMouse, Open, Release, Wait
 from screenpy_selenium.questions import Text
 
-from ..user_interface.drag_and_drop import (
+from screenpy_examples.screenpy_selenium.the_internet.user_interface.drag_and_drop import \
+    (
     FIRST_DRAGGABLE_BOX,
     SECOND_DRAGGABLE_BOX,
     URL,
