+++ screenpy_examples/screenpy_selenium/the_internet/features_unittest/test_dropdowns.py
@@ -15,7 +15,10 @@ from screenpy_selenium.abilities import BrowseTheWeb
 from screenpy_selenium.actions import Open, Select
 from screenpy_selenium.questions import Selected
 
-from ..user_interface.dropdown import THE_DROPDOWN, URL
+from screenpy_examples.screenpy_selenium.the_internet.user_interface.dropdown import (
+    THE_DROPDOWN,
+    URL,
+)
 
 
 class TestDropdowns(unittest.TestCase):
