+++ screenpy_examples/screenpy_selenium/the_internet/features_unittest/test_multiple_windows.py
@@ -13,7 +13,7 @@ from screenpy_selenium.abilities import BrowseTheWeb
 from screenpy_selenium.actions import Click, Open, SwitchToTab
 from screenpy_selenium.questions import BrowserURL, Text
 
-from ..user_interface.multiple_windows import (
+from screenpy_examples.screenpy_selenium.the_internet.user_interface.multiple_windows import (
     CLICK_HERE_LINK,
     HEADER_MESSAGE,
     URL,
